src/slm_finetuner.py
import json
import torch
from transformers import AutoProcessor, AutoModelForVision2Seq
from peft import LoraConfig, get_peft_model
import google.cloud.aiplatform as aip

class SLMFinetuner:
    """Fine-tunes a Small Vision-Language Model (SLM) for Manufacturing Detection."""

    def __init__(self, model_id="microsoft/Phi-3-vision-128k-instruct"):
        self.model_id = model_id
        self.processor = AutoProcessor.from_pretrained(model_id, trust_remote_code=True)
        self.model = None

    def load_model_for_training(self):
        """Prepares model with LoRA/QLoRA for memory efficiency."""
        print(f"Loading {self.model_id} for fine-tuning...")
        self.model = AutoModelForVision2Seq.from_pretrained(
            self.model_id, 
            trust_remote_code=True,
            torch_dtype=torch.float16,
            device_map="auto"
        )
        
        # LoRA Config for vision-language tasks
        config = LoraConfig(
            r=16,
            lora_alpha=32,
            target_modules=["q_proj", "v_proj", "k_proj", "out_proj"],
            lora_dropout=0.05,
            bias="none",
            task_type="CAUSAL_LM"
        )
        self.model = get_peft_model(self.model, config)
        print("Model wrapped with LoRA. Ready for Training.")

    def preprocess_dataset(self, dataset_path):
        """Converts raw manufacturing images/annotations into SFT format."""
        # This would handle the conversion of YOLO/JSON data to VLM-SFT formats
        pass

    def run_gcp_ml_job(self, workspace_params):
        """Schedules the fine-tuning job on a GCP Vertex AI GPU instance."""
        try:
            aip.init(
                project=workspace_params.get('project_id'),
                location=workspace_params.get('region', 'us-central1')
            )
            # Define a GCP Vertex AI Job configuration here
            print("GCP Vertex AI Client ready. Submit training job via Vertex AI SDK.")
        except Exception as e:
            print(f"Failed to connect to GCP: {e}")

if __name__ == "__main__":
    tuner = SLMFinetuner()
    # Note: Training usually requires A100/H100/V100 on GCP
    # tuner.load_model_for_training()
    print("SLM Fine-tuning setup generated.")
    print("Check docs/GENAI_GCP_ARCHITECTURE.md for infra setup.")
