requirements.txt
pandas
numpy
scikit-learn
flask
joblib
matplotlib
seaborn
google-cloud-aiplatform
google-cloud-storage
google-cloud-iot
google-cloud-firestore
langchain                    
langchain-google-vertexai      # GenAI Reasoning
ultralytics                    # Real-time base detection
transformers                   # SLM (Phi-3/Qwen) handling
peft                           # LoRA/QLoRA for fine-tuning
bitsandbytes                   # 4-bit quant
accelerate                     # multi-gpu training
pydantic                       # Structured event validation
opencv-python
imutils
python-dotenv